linux/src/requirements.txt
pyudev
pystray
pillow
PyGObject
